refactor: rearrange code
- move out unrelated code from init.lua to dedicated files
- add strategy.lua with dap/neotest strategy configuration details
- add results.lua which is managing test results
- refactor framework files into directory structure with init.lua files
  as entrypoints
